Extract flair logic into src/core/flair.ts
Move approval flair config, template validation, viewer flair
snapshots, and verification flair checks into a dedicated module.
Typecheck, 167 tests, and build all pass.
